requirements.txt
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.0.1
numpy==1.25.2
qrcode[pil]==7.4.2
pyttsx3==2.90
# Additional packages for stage detection
opencv-contrib-python==4.8.1.78
# Additional packages for photo processing
requests==2.31.0
scikit-image==0.21.0
Flask==2.3.3
Flask-Cors==4.0.0